feat: add settings model traits
The SettingsModel trait allows users to define their own settings models
to be made available within the Bottlerocket API.

Implementors of SettingsModel are wrapped in a BottlerocketSetting
struct, which implements Model: a type-erased analogue to the
SettingsModel trait. This allows settings extensions to store a list of
references to trait objects offering the behavior defined by
SettingsModels.