proto1: parse arguments as JSON from a file
It's possible for proto1's command line interface to exceed the OS's
maximum argument length (128 KiB) if the JSON values passed into the
command are particularly large.

So instead of getting arguments from the command line, we parse them as
an equivalent JSON structure from a provided file. By default, this is
from stdin, but this can be overridden with the --input-file flag.

Signed-off-by: Sam Berning <bernings@amazon.com>